+++ internal/config/loader.go
@@ -22,11 +22,3 @@ func LoadConfig(path string) (*Config, error) {
 	}
 	return cfg, nil
 }
-
-func SaveConfig(cfg *Config, path string) error {
-	data, err := toml.Marshal(cfg)
-	if err != nil {
-		return err
-	}
-	return os.WriteFile(path, data, 0644)
-}
