+++ internal/constant/max_short_length.go
@@ -1,0 +1,3 @@
+package constant
+
+const MaxShortLen = 20;
