+++ internal/domain/command_entry.go
@@ -1,0 +1,7 @@
+package domain
+
+type CommandEntry struct {
+	Prefix  string
+	Short   string
+	Command string
+}
