+++ internal/handler/exec.go
@@ -16,12 +16,12 @@ func ExecCommand(configPath, prefix, short string) error {
 
 	cmds, exists := cfg.Commands[prefix]
 	if !exists {
-		return fmt.Errorf("command not found: %s/%s", prefix, short)
+		return fmt.Errorf("command not found: %s %s", prefix, short)
 	}
 
 	command, ok := cmds[short]
 	if !ok {
-		return fmt.Errorf("command not found: %s/%s", prefix, short)
+		return fmt.Errorf("command not found: %s %s", prefix, short)
 	}
 
 	execCmd := exec.Command("sh", "-c", command)
